Struct openstack::auth::Identity

pub struct Identity { /* fields omitted */ }

Authentication method factory using Identity API V3.

Methods

impl Identity

pub fn auth_url(&self) -> &Url

Get a reference to the auth URL.

pub fn new<U>(auth_url: U) -> Result<Identity, UrlError> where
U: IntoUrl,

Create a password authentication against the given Identity service.

pub fn new_with_region<U>(
auth_url: U,
region: String
) -> Result<Identity, UrlError> where
U: IntoUrl,

Create a password authentication against the given Identity service.

pub fn new_with_client<U>(
auth_url: U,
client: Client
) -> Result<Identity, UrlError> where
U: IntoUrl,

Create a password authentication against the given Identity service.

pub fn with_user<S1, S2, S3>(
self,
user_name: S1,
password: S2,
domain_name: S3
) -> Identity where
S1: Into<String>,
S2: Into<String>,
S3: Into<String>,

Add authentication based on user name and password.

pub fn with_project_scope<S1, S2>(
self,
project_name: S1,
domain_name: S2
) -> Identity where
S1: Into<String>,
S2: Into<String>,

Request a token scoped to the given project.

pub fn create(self) -> Result<PasswordAuth>

Create an authentication method based on provided information.