Struct openstack::auth::Password

pub struct Password { /* fields omitted */ }

Password authentication using Identity API V3.

For any Identity authentication you need to know auth_url, which is an authentication endpoint of the Identity service. For the Password authentication you also need:

User name and password.

Domain of the user.

Name of the project to use.

Domain of the project.

Note: currently only names are supported for user, user domain and project domain. ID support is coming later.

Start with creating a Password object using new, then add a project scope with with_project_scope:

let auth = osauth::identity::Password::new(
    "https://cloud.local/identity",
    "admin",
    "pa$$w0rd",
    "Default"
)
.expect("Invalid auth_url")
.with_project_scope(IdOrName::from_name("project1"), IdOrName::from_id("default"));

let session = osauth::Session::new(auth);

If your cloud has several regions, pick one using with_region:

use osauth::identity::IdOrName;

let scope = osauth::identity::Scope::Project {
    project: IdOrName::from_name("project1"),
    domain: Some(IdOrName::from_id("default")),
};
let auth = osauth::identity::Password::new(
    "https://cloud.local/identity",
    "admin",
    "pa$$w0rd",
    "Default"
)
.expect("Invalid auth_url")
.with_scope(scope)
.with_region("US-East");

let session = osauth::Session::new(auth);

By default, the public endpoint interface is used. f you would prefer to default to another one, you can set it with with_default_endpoint_interface.

use osauth::identity::IdOrName;

let scope = osauth::identity::Scope::Project {
    project: IdOrName::from_name("project1"),
    domain: Some(IdOrName::from_id("default")),
};
let auth = osauth::identity::Password::new(
    "https://cloud.local/identity",
    "admin",
    "pa$$w0rd",
    "Default"
)
.expect("Invalid auth_url")
.with_scope(scope)
.with_default_endpoint_interface(osauth::InterfaceType::Internal);

The authentication token is cached while it's still valid or until refresh is called. Clones of a Password also start with an empty cache.

Implementations

impl Password

pub fn new<U, S1, S2, S3>(
auth_url: U,
user_name: S1,
password: S2,
user_domain_name: S3
) -> Result<Password, Error> where
S1: Into<String>,
S2: Into<String>,
S3: Into<String>,
U: IntoUrl,

Create a password authentication.

pub fn new_with_client<U, S1, S2, S3>(
auth_url: U,
client: Client,
user_name: S1,
password: S2,
user_domain_name: S3
) -> Result<Password, Error> where
S1: Into<String>,
S2: Into<String>,
S3: Into<String>,
U: IntoUrl,

Create a password authentication with the provided HTTP client.

pub fn endpoint_filters(&self) -> &EndpointFilters

Endpoint filters.

pub fn endpoint_filters_mut(&mut self) -> &mut EndpointFilters

Mutable endpoint filters.

pub fn set_default_endpoint_interface(
&mut self,
endpoint_interface: InterfaceType
)

Set the default endpoint interface to use.

pub fn set_endpoint_filters(&mut self, filters: EndpointFilters)

Set endpoint filters.

pub fn set_region<S>(&mut self, region: S) where
S: Into<String>,

👎 Deprecated since 0.3.0:
Use set_filters or filters_mut

Set a region for this authentication method.

pub fn set_project_scope(
&mut self,
project: IdOrName,
domain: impl Into<Option<IdOrName>>
)

Scope authentication to the given project.

A convenience wrapper around set_scope.

pub fn set_scope(&mut self, scope: Scope)

Add a scope to the authentication.

This is required in the most cases.

pub fn with_default_endpoint_interface(
self,
endpoint_interface: InterfaceType
) -> Password

Convert this authentication into one using the given endpoint interface.

pub fn with_endpoint_filters(self, filters: EndpointFilters) -> Password

Add endpoint filters.

pub fn with_project_scope(
self,
project: IdOrName,
domain: impl Into<Option<IdOrName>>
) -> Password

Scope authentication to the given project.

A convenience wrapper around with_scope.

pub fn with_region<S>(self, region: S) -> Password where
S: Into<String>,

Set a region for this authentication method.

pub fn with_scope(self, scope: Scope) -> Password

Add a scope to the authentication.

pub fn user(&self) -> &IdOrName

User name or ID.

pub fn project(&self) -> Option<&IdOrName>

Project name or ID (if project scoped).