Trait openstack::auth::AuthType

pub trait AuthType: Send + Sync + Debug {
    fn get_endpoint(
        &self, 
        service_type: String, 
        endpoint_interface: Option<String>
    ) -> Box<dyn Future<Item = Url, Error = Error> + 'static + Send>;

    fn request(
        &self, 
        method: Method, 
        url: Url
    ) -> Box<dyn Future<Item = RequestBuilder, Error = Error> + 'static + Send>;

    fn refresh(
        &self
    ) -> Box<dyn Future<Item = (), Error = Error> + 'static + Send>;

    fn region(&self) -> Option<String> { ... }
}

Trait for an authentication type.

An OpenStack authentication type is expected to be able to:

get an authentication token to use when accessing services,

get an endpoint URL for the given service type.

An authentication type should cache the token as long as it's valid.

Required methods

fn get_endpoint(
&self,
service_type: String,
endpoint_interface: Option<String>
) -> Box<dyn Future<Item = Url, Error = Error> + 'static + Send>

Get a URL for the requested service.

fn request(
&self,
method: Method,
url: Url
) -> Box<dyn Future<Item = RequestBuilder, Error = Error> + 'static + Send>

Create an authenticated request.

fn refresh(&self) -> Box<dyn Future<Item = (), Error = Error> + 'static + Send>

Refresh the authentication (renew the token, etc).

Loading content...
Provided methods

fn region(&self) -> Option<String>

Region used with this authentication (if any).

Loading content...