Struct openstack::compute::FlavorQuery

pub struct FlavorQuery { /* fields omitted */ }

A query to server list.

Methods

impl FlavorQuery

pub fn with_marker<T: Into<String>>(self, marker: T) -> Self

Add marker to the request.

Using this disables automatic pagination.

pub fn with_limit(self, limit: usize) -> Self

Add limit to the request.

Using this disables automatic pagination.

pub fn into_iter(self) -> ResourceIterator<FlavorSummary>

Convert this query into an iterator executing the request.

This iterator yields only FlavorSummary objects, containing IDs and names. Use into_iter_detailed for full Flavor objects.

Returns a FallibleIterator, which is an iterator with each next call returning a Result.

Note that no requests are done until you start iterating.

pub fn into_iter_detailed(self) -> ResourceIterator<Flavor>

Convert this query into an iterator executing the request.

This iterator yields full Flavor objects. If you only need IDs and/or names, use into_iter to save bandwidth.

Returns a FallibleIterator, which is an iterator with each next call returning a Result.

Note that no requests are done until you start iterating.

pub fn all(self) -> Result<Vec<FlavorSummary>>

Execute this request and return all results.

A convenience shortcut for self.into_iter().collect().

pub fn one(self) -> Result<FlavorSummary>

Return one and exactly one result.

Fails with ResourceNotFound if the query produces no results and with TooManyItems if the query produces more than one result.