Struct openstack::compute::Server

pub struct Server { /* fields omitted */ }

Structure representing a single server.

Methods

impl Server

pub fn access_ipv4(&self) -> Option<Ipv4Addr>

IPv4 address to access the server (if provided).

pub fn access_ipv6(&self) -> Option<Ipv6Addr>

IPv6 address to access the server (if provided).

pub fn addresses(&self) -> &HashMap<String, Vec<ServerAddress>>

Addresses (floating and fixed) associated with the server.

pub fn availability_zone(&self) -> &String

Availability zone.

pub fn created_at(&self) -> DateTime<FixedOffset>

Creation date and time.

pub fn description(&self) -> &Option<String>

Server description.

pub fn flavor(&self) -> &ServerFlavor

Flavor information used to create this server.

pub fn floating_ip(&self) -> Option<IpAddr>

Find a floating IP, if it exists.

If multiple floating IPs exist, the first is returned.

pub fn has_config_drive(&self) -> bool

Whether the server was created with a config drive.

pub fn has_image(&self) -> bool

Whether the server has an image.

May return false if the server was created from a volume.

pub fn id(&self) -> &String

Server unique ID.

pub fn image(&self) -> Result<Image>

Fetch the associated image.

Fails with ResourceNotFound if the server does not have an image.

pub fn image_id(&self) -> Option<&String>

Get a reference to the image.

May be None if the server was created from a volume.

pub fn key_pair(&self) -> Result<KeyPair>

Fetch the key pair used for the server.

pub fn key_pair_name(&self) -> &Option<String>

Name of a key pair used with this server (if any).

pub fn name(&self) -> &String

Server name.

pub fn metadata(&self) -> &HashMap<String, String>

Metadata associated with the server.

pub fn power_state(&self) -> ServerPowerState

Server power state.

pub fn status(&self) -> ServerStatus

Server status.

pub fn updated_at(&self) -> DateTime<FixedOffset>

Last update date and time.

pub fn delete(self) -> Result<DeletionWaiter<Server>>

Delete the server.

pub fn reboot<'server>(
&'server mut self,
reboot_type: RebootType
) -> Result<ServerStatusWaiter<'server>>

Reboot the server.

pub fn start<'server>(&'server mut self) -> Result<ServerStatusWaiter<'server>>

Start the server, optionally wait for it to be active.

pub fn stop<'server>(&'server mut self) -> Result<ServerStatusWaiter<'server>>

Stop the server, optionally wait for it to be powered off.