Struct openstack::network::NetworkQuery

pub struct NetworkQuery { /* fields omitted */ }

A query to network list.

Methods

impl NetworkQuery

pub fn with_marker<T: Into<String>>(self, marker: T) -> Self

Add marker to the request.

Using this disables automatic pagination.

pub fn with_limit(self, limit: usize) -> Self

Add limit to the request.

Using this disables automatic pagination.

pub fn sort_by(self, sort: Sort<NetworkSortKey>) -> Self

Add sorting to the request.

pub fn with_name<T: Into<String>>(self, value: T) -> Self

Filter by network name (a database regular expression).

pub fn into_iter(self) -> ResourceIterator<Network>

Convert this query into an iterator executing the request.

Returns a FallibleIterator, which is an iterator with each next call returning a Result.

Note that no requests are done until you start iterating.

pub fn all(self) -> Result<Vec<Network>>

Execute this request and return all results.

A convenience shortcut for self.into_iter().collect().

pub fn one(self) -> Result<Network>

Return one and exactly one result.

Fails with ResourceNotFound if the query produces no results and with TooManyItems if the query produces more than one result.