Struct openstack::network::SubnetQuery

pub struct SubnetQuery { /* fields omitted */ }

A query to subnet list.

Methods

impl SubnetQuery

pub fn with_marker<T: Into<String>>(self, marker: T) -> Self

Add marker to the request.

Using this disables automatic pagination.

pub fn with_limit(self, limit: usize) -> Self

Add limit to the request.

Using this disables automatic pagination.

pub fn sort_by(self, sort: Sort<SubnetSortKey>) -> Self

Add sorting to the request.

pub fn set_cidr(&mut self, value: IpNet)

Filter by CIDR.

pub fn with_cidr(self, value: IpNet) -> Self

Filter by CIDR.

pub fn set_description<T: Into<String>>(&mut self, value: T)

Filter by description.

pub fn with_description<T: Into<String>>(self, value: T) -> Self

Filter by description.

pub fn set_dhcp_enabled(&mut self, value: bool)

Filter by whether DHCP is enabled.

pub fn with_dhcp_enabled(self, value: bool) -> Self

Filter by whether DHCP is enabled.

pub fn set_gateway_ip(&mut self, value: IpAddr)

Filter by gateway IP.

pub fn with_gateway_ip(self, value: IpAddr) -> Self

Filter by gateway IP.

pub fn set_ipv6_address_mode(&mut self, value: Ipv6Mode)

Filter by IPv6 address assignment mode.

pub fn with_ipv6_address_mode(self, value: Ipv6Mode) -> Self

Filter by IPv6 address assignment mode.

pub fn set_ipv6_router_advertisement_mode(&mut self, value: Ipv6Mode)

Filter by IPv6 router advertisement mode.

pub fn with_ipv6_router_advertisement(self, value: Ipv6Mode) -> Self

Filter by IPv6 router advertisement mode.

pub fn set_name<T: Into<String>>(&mut self, value: T)

Filter by subnet name.

pub fn with_name<T: Into<String>>(self, value: T) -> Self

Filter by subnet name.

pub fn set_network<N: Into<NetworkRef>>(&mut self, value: N)

Filter by network.

Warning

Due to architectural limitations, names do not work here.

pub fn with_network<N: Into<NetworkRef>>(self, value: N) -> Self

Filter by network.

Warning

Due to architectural limitations, names do not work here.

pub fn into_iter(self) -> ResourceIterator<Subnet>

Convert this query into an iterator executing the request.

Returns a FallibleIterator, which is an iterator with each next call returning a Result.

Note that no requests are done until you start iterating.

pub fn all(self) -> Result<Vec<Subnet>>

Execute this request and return all results.

A convenience shortcut for self.into_iter().collect().

pub fn one(self) -> Result<Subnet>

Return one and exactly one result.

Fails with ResourceNotFound if the query produces no results and with TooManyItems if the query produces more than one result.