Struct openstack::session::RequestBuilder

pub struct RequestBuilder { /* fields omitted */ }

An HTTP request builder.

This is a thin wrapper around reqwest's RequestBuilder with error handling.

Methods

impl RequestBuilder

pub fn new(inner: ReqwestRB) -> RequestBuilder

Create a RequestBuilder by wrapping a reqwest's one.

pub fn inner_mut(&mut self) -> &mut ReqwestRB

Access to the inner object.

pub fn into_inner(self) -> ReqwestRB

Take the inner object out.

pub fn header<H: Header>(&mut self, header: H) -> &mut RequestBuilder

Add a Header to this Request.

pub fn headers(&mut self, headers: Headers) -> &mut RequestBuilder

Add a set of Headers to the existing ones on this Request.

pub fn body<T: Into<Body>>(&mut self, body: T) -> &mut RequestBuilder

Set the request body.

pub fn query<T: Serialize>(&mut self, query: &T) -> &mut RequestBuilder

Modify the query string of the URL.

pub fn json<T: Serialize>(&mut self, json: &T) -> &mut RequestBuilder

Send a JSON body.

pub fn send(&mut self) -> Result<Response>

Construct the Request and sends it the target URL, returning a Response.

pub fn receive_json<T: DeserializeOwned>(&mut self) -> Result<T>

Construct the Request, send it and receive a JSON.