Crate openstack

OpenStack Client in Rust.

The goal of this project is to provide a simple API for working with OpenStack clouds.

Authentication

High-level API

API-specific notes:

Compute API support

Modules

auth
Authentication modules.

compute
Compute API implementation bits.

service
Generic API bits for implementing new services.

session
Session structure definition.

Structs

ApiVersion
API version (major, minor).

Cloud
OpenStack cloud API.

Error
Error from an OpenStack call.

Enums

ApiVersionRequest
A request for negotiating an API version.

ErrorKind
Kind of an error.

Sort
Sorting request.

Traits

Waiter
Trait representing a waiter for some asynchronous action to finish.

Type Definitions

Result
Result of an OpenStack call.