Crate openstack

OpenStack SDK in Rust.

The goal of this project is to provide a simple API for working with OpenStack clouds.

Usage

Start with authentication, then create a Cloud object and use it for OpenStack API calls.

Example

Get authentication parameters from the environment and get UUIDs of all servers.

extern crate openstack;

fn get_server_uuids() -> openstack::Result<Vec<String>> {
    let os = openstack::Cloud::from_env()?;
    let servers = os.list_servers()?;
    Ok(servers.into_iter().map(|server| server.id().clone()).collect())
}

See Cloud struct for more examples.

Modules

auth
Authentication modules.

common
Types and traits shared by all API parts.

compute
Compute API implementation bits.

image
Image API implementation bits.

network
Network API implementation bits.

session
Session structure definition.

Structs

Cloud
OpenStack cloud API.

Error
Error from an OpenStack call.

Enums

ErrorKind
Kind of an error.

Sort
Sorting request.

Traits

Refresh
Trait representing something that can be refreshed.

Type Definitions

Result
Result of an OpenStack call.