Trait openstack::common::ResourceQuery

pub trait ResourceQuery {
    type Item;

    const DEFAULT_LIMIT: usize;

    fn can_paginate(&self) -> Result<bool>;

    fn extract_marker(&self, resource: &Self::Item) -> String;

    fn fetch_chunk(
        &self, 
        limit: Option<usize>, 
        marker: Option<String>
    ) -> Result<Vec<Self::Item>>;

    fn validate(&mut self) -> Result<()> { ... }
}

A query for resources.

This is a low-level trait that should not be used directly.

Associated Types

type Item

Item type.

Loading content...
Associated Constants

const DEFAULT_LIMIT: usize

Default limit to use with this query.

Loading content...
Required methods

fn can_paginate(&self) -> Result<bool>

Whether pagination is supported for this query.

fn extract_marker(&self, resource: &Self::Item) -> String

Extract a marker from a resource.

fn fetch_chunk(
&self,
limit: Option<usize>,
marker: Option<String>
) -> Result<Vec<Self::Item>>

Get a chunk of resources.

Loading content...
Provided methods

fn validate(&mut self) -> Result<()>

Validate the query before the first execution.

This call may modify internal representation of the query, so changing the query after calling it may cause undesired side effects.

Loading content...