Struct openstack::compute::NewServer

pub struct NewServer { /* fields omitted */ }

A request to create a server.

Methods

impl NewServer

pub fn create(self) -> Result<ServerCreationWaiter>

Request creation of the server.

pub fn add_fixed_ip(&mut self, fixed_ip: Ipv4Addr)

Add a virtual NIC with given fixed IP to the new server.

A shorthand for add_nic.

pub fn add_network<N>(&mut self, network: N) where
N: Into<NetworkRef>,

Add a virtual NIC from this network to the new server.

A shorthand for add_nic.

pub fn add_nic(&mut self, nic: ServerNIC)

Add a virtual NIC to the new server.

pub fn add_port<P>(&mut self, port: P) where
P: Into<PortRef>,

Add a virtual NIC with this port to the new server.

A shorthand for add_nic.

pub fn set_image<I>(&mut self, image: I) where
I: Into<ImageRef>,

Use this image as a source for the new server.

pub fn with_fixed_ip(self, fixed_ip: Ipv4Addr) -> NewServer

Add a virtual NIC with given fixed IP to the new server.

pub fn with_image<I>(self, image: I) -> NewServer where
I: Into<ImageRef>,

Use this image as a source for the new server.

pub fn with_network<N>(self, network: N) -> NewServer where
N: Into<NetworkRef>,

Add a virtual NIC from this network to the new server.

pub fn with_port<P>(self, port: P) -> NewServer where
P: Into<PortRef>,

Add a virtual NIC with this port to the new server.