Struct openstack::compute::Server

pub struct Server { /* fields omitted */ }

Structure representing a single server.

Methods

impl Server

pub fn access_ipv4(&self) -> Option<Ipv4Addr>

Get the IPv4 address.

If not None, this address should be used to access the server instead of one from the addresses method.

pub fn access_ipv6(&self) -> Option<Ipv6Addr>

Get the IPv6 address.

If not None, this address should be used to access the server instead of one from the addresses method.

pub fn addresses(&self) -> &HashMap<String, Vec<ServerAddress>>

Get a reference to associated addresses.

pub fn availability_zone(&self) -> &String

Get a reference to the availability zone.

pub fn created_at(&self) -> DateTime<FixedOffset>

Get the creation date and time.

pub fn floating_ip(&self) -> Option<IpAddr>

Find a floating IP, if it exists.

If multiple floating IPs exist, the first is returned.

pub fn id(&self) -> &String

Get a reference to server unique ID.

pub fn image_id(&self) -> Option<&String>

Get a reference to the image.

May be None if the server was created from a volume.

pub fn has_image(&self) -> bool

Whether the server has an image.

May return false if the server was created from a volume.

pub fn name(&self) -> &String

Get a reference to server name.

pub fn power_state(&self) -> ServerPowerState

Get server power state.

pub fn status(&self) -> ServerStatus

Get server status.

pub fn updated_at(&self) -> DateTime<FixedOffset>

Get the last update date and time.

pub fn delete(self) -> Result<DeletionWaiter<Server>>

Delete the server.

pub fn reboot<'server>(
&'server mut self,
reboot_type: RebootType
) -> Result<ServerStatusWaiter<'server>>

Reboot the server.

pub fn start<'server>(&'server mut self) -> Result<ServerStatusWaiter<'server>>

Start the server, optionally wait for it to be active.

pub fn stop<'server>(&'server mut self) -> Result<ServerStatusWaiter<'server>>

Stop the server, optionally wait for it to be powered off.