Struct openstack::compute::ServerQuery

pub struct ServerQuery<'session> { /* fields omitted */ }

A query to server list.

Methods

impl<'session> ServerQuery<'session>

pub fn with_marker<T: Into<String>>(self, marker: T) -> Self

Add marker to the request.

Using this disables automatic pagination.

pub fn with_limit(self, limit: usize) -> Self

Add limit to the request.

Using this disables automatic pagination.

pub fn sort_by(self, sort: Sort<ServerSortKey>) -> Self

Add sorting to the request.

pub fn with_access_ip_v4(self, value: Ipv4Addr) -> Self

Filter by IPv4 address that should be used to access the server.

pub fn with_access_ip_v6(self, value: Ipv6Addr) -> Self

Filter by IPv6 address that should be used to access the server.

pub fn with_availability_zone<T: Into<String>>(self, value: T) -> Self

Filter by availability zone.

pub fn with_flavor<T: ToFlavorId>(self, value: T) -> Self

Filter by flavor.

pub fn with_hostname<T: Into<String>>(self, value: T) -> Self

Filter by host name.

pub fn with_image<T: ToImageId>(self, value: T) -> Self

Filter by image ID.

pub fn with_ip_v4(self, value: Ipv4Addr) -> Self

Filter by an IPv4 address.

pub fn with_ip_v6(self, value: Ipv6Addr) -> Self

Filter by an IPv6 address.

pub fn with_name<T: Into<String>>(self, value: T) -> Self

Filter by server name (a database regular expression).

pub fn with_power_state<T: Into<String>>(self, value: T) -> Self

Filter by power state.

pub fn with_project_id<T: Into<String>>(self, value: T) -> Self

Filter by project ID (also commonly known as tenant ID).

pub fn with_status<T: Into<String>>(self, value: T) -> Self

Filter by server status.

pub fn with_user_id<T: Into<String>>(self, value: T) -> Self

Filter by user ID.

pub fn into_iter(self) -> ResourceIterator<'session, ServerSummary<'session>>

Convert this query into an iterator executing the request.

This iterator yields only ServerSummary objects, containing IDs and names. Use into_iter_detailed for full Server objects.

Returns a FallibleIterator, which is an iterator with each next call returning a Result.

Note that no requests are done until you start iterating.

pub fn into_iter_detailed(self) -> ResourceIterator<'session, Server<'session>>

Convert this query into an iterator executing the request.

This iterator yields full Server objects. If you only need IDs and/or names, use into_iter to save bandwidth.

Returns a FallibleIterator, which is an iterator with each next call returning a Result.

Note that no requests are done until you start iterating.

pub fn all(self) -> Result<Vec<ServerSummary<'session>>>

Execute this request and return all results.

A convenience shortcut for self.into_iter().collect().

pub fn one(self) -> Result<ServerSummary<'session>>

Return one and exactly one result.

Fails with ResourceNotFound if the query produces no results and with TooManyItems if the query produces more than one result.