Struct openstack::compute::ServerQuery

pub struct ServerQuery { /* fields omitted */ }

A query to server list.

Methods

impl ServerQuery

pub fn with_marker<T: Into<String>>(self, marker: T) -> Self

Add marker to the request.

Using this disables automatic pagination.

pub fn with_limit(self, limit: usize) -> Self

Add limit to the request.

Using this disables automatic pagination.

pub fn sort_by(self, sort: Sort<ServerSortKey>) -> Self

Add sorting to the request.

pub fn set_access_ip_v4<T: Into<Ipv4Addr>>(&mut self, value: T)

Filter by IPv4 address that should be used to access the server.

pub fn with_access_ip_v4<T: Into<Ipv4Addr>>(self, value: T) -> Self

Filter by IPv4 address that should be used to access the server.

pub fn set_access_ip_v6<T: Into<Ipv6Addr>>(&mut self, value: T)

Filter by IPv6 address that should be used to access the server.

pub fn with_access_ip_v6<T: Into<Ipv6Addr>>(self, value: T) -> Self

Filter by IPv6 address that should be used to access the server.

pub fn set_availability_zone<T: Into<String>>(&mut self, value: T)

Filter by availability zone.

pub fn with_availability_zone<T: Into<String>>(self, value: T) -> Self

Filter by availability zone.

pub fn set_flavor<T: Into<FlavorRef>>(&mut self, value: T)

Filter by flavor.

pub fn with_flavor<T: Into<FlavorRef>>(self, value: T) -> Self

Filter by flavor.

pub fn set_hostname<T: Into<String>>(&mut self, value: T)

Filter by host name.

pub fn with_hostname<T: Into<String>>(self, value: T) -> Self

Filter by host name.

pub fn set_image<T: Into<ImageRef>>(&mut self, value: T)

Filter by image used to build the server.

pub fn with_image<T: Into<ImageRef>>(self, value: T) -> Self

Filter by image used to build the server.

pub fn set_ip_v4<T: Into<Ipv4Addr>>(&mut self, value: T)

Filter by an IPv4 address.

pub fn with_ip_v4<T: Into<Ipv4Addr>>(self, value: T) -> Self

Filter by an IPv4 address.

pub fn set_ip_v6<T: Into<Ipv6Addr>>(&mut self, value: T)

Filter by an IPv6 address.

pub fn with_ip_v6<T: Into<Ipv6Addr>>(self, value: T) -> Self

Filter by an IPv6 address.

pub fn set_name<T: Into<String>>(&mut self, value: T)

Filter by name.

pub fn with_name<T: Into<String>>(self, value: T) -> Self

Filter by name.

pub fn set_project<T: Into<ProjectRef>>(&mut self, value: T)

Filter by project (also commonly known as tenant).

pub fn with_project<T: Into<ProjectRef>>(self, value: T) -> Self

Filter by project (also commonly known as tenant).

pub fn set_status<T: Into<ServerStatus>>(&mut self, value: T)

Filter by server status.

pub fn with_status<T: Into<ServerStatus>>(self, value: T) -> Self

Filter by server status.

pub fn set_user<T: Into<UserRef>>(&mut self, value: T)

Filter by user.

pub fn with_user<T: Into<UserRef>>(self, value: T) -> Self

Filter by user.

pub fn detailed(self) -> DetailedServerQuery

Convert this query into a detailed query.

pub fn into_iter(self) -> ResourceIterator<ServerQuery>

Convert this query into an iterator executing the request.

This iterator yields only ServerSummary objects, containing IDs and names. Use into_iter_detailed for full Server objects.

Returns a FallibleIterator, which is an iterator with each next call returning a Result.

Note that no requests are done until you start iterating.

pub fn all(self) -> Result<Vec<ServerSummary>>

Execute this request and return all results.

A convenience shortcut for self.into_iter().collect().

pub fn one(self) -> Result<ServerSummary>

Return one and exactly one result.

Fails with ResourceNotFound if the query produces no results and with TooManyItems if the query produces more than one result.