Struct openstack::session::Session

pub struct Session { /* fields omitted */ }

An OpenStack API session.

The session object serves as a wrapper around an HTTP(s) client, handling authentication, accessing the service catalog and token refresh.

The session object also owns region and endpoint interface to use.

Finally, the session object is responsible for API version negotiation.

Methods

impl Session

pub fn new<Auth: AuthMethod + 'static>(auth_method: Auth) -> Session

Create a new session with a given authentication plugin.

The resulting session will use the default endpoint interface (usually, public) and the first available region.

pub fn with_endpoint_interface<S>(self, endpoint_interface: S) -> Session where
S: Into<String>,

Convert this session into one using the given endpoint interface.

Negotiated API versions are kept in the new object.

pub fn auth_method(&self) -> &AuthMethod

Get a reference to the authentication method in use.

pub fn api_version<Srv: ServiceType>(&self) -> Option<ApiVersion>

Get an API version used for given service.

pub fn get_service_info<Srv>(&self) -> Result<ServiceInfo> where
Srv: ServiceType,

Get service info for the given service.

pub fn get_endpoint<Srv: ServiceType>(&self, path: &[&str]) -> Result<Url>

Construct and endpoint for the given service from the path.

pub fn request<Srv: ServiceType>(
&self,
method: Method,
path: &[&str]
) -> Result<RequestBuilder>

Make an HTTP request to the given service.

pub fn negotiate_api_version<Srv>(
&mut self,
requested: ApiVersionRequest
) -> Result<ApiVersion> where
Srv: ServiceType + ApiVersioning,

Negotiate an API version with the service.

Negotiation is based on version information returned from the root endpoint. If no minimum version is returned, the current version is assumed to be the only supported version.

The resulting API version is cached for this session.